refactor: explicitly export public API from entry points
