+++ packages/rainbowkit/src/components/ExplorerLink/ExplorerLink.tsx
@@ -1,14 +1,14 @@
 import clsx from 'clsx'
 import React from 'react'
-import { chainIDToExplorer } from '../../utils'
+import { chainIDToExplorer } from '../../utils/convert'
 import { ExplorerLinkClassName } from './ExplorerLink.css'
 
-export type ExplorerProps = { chainId?: number; address: string; explorerUrl?: string } & React.DetailedHTMLProps<
+export type ExplorerLinkProps = { chainId?: number; address: string; explorerUrl?: string } & React.DetailedHTMLProps<
   React.AnchorHTMLAttributes<HTMLAnchorElement>,
   HTMLAnchorElement
 >
 
-export const ExplorerLink = ({ chainId = 1, address, explorerUrl, className, ...props }: ExplorerProps) => (
+export const ExplorerLink = ({ chainId = 1, address, explorerUrl, className, ...props }: ExplorerLinkProps) => (
   <a
     href={`${explorerUrl || chainIDToExplorer(chainId).url}/address/${address}`}
     target="_blank"
