+++ packages/rainbowkit/src/components/Modal/Modal.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState, Dispatch } from 'react'
 import type { UserRejectedRequestError } from '@web3-react/injected-connector'
 import type { UnsupportedChainIdError } from '@web3-react/core'
 import clsx from 'clsx'
-import { getWalletInfo, Wallet } from '../../utils'
+import { getWalletInfo, Wallet } from '../../utils/wallets'
 import * as styles from './Modal.css'
 import { CloseIcon, NextIcon } from './icons'
 import { Box } from '../Box/Box'
