+++ packages/rainbowkit/src/components/Profile/Profile.tsx
@@ -1,5 +1,8 @@
 import React, { Dispatch, useMemo, useRef } from 'react'
-import { useENSWithAvatar, useOnClickOutside, useToggle, useWalletModal, UseWalletModalOptions } from '../../hooks'
+import { useENSWithAvatar } from '../../hooks/useENSWithAvatar'
+import { useOnClickOutside } from '../../hooks/useOnClickOutside'
+import { useToggle } from '../../hooks/useToggle'
+import { useWalletModal, UseWalletModalOptions } from '../../hooks/useWalletModal'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { Badge } from '../Badge/Badge'
 import { WalletDropdown, WalletDropdownProps } from '../WalletDropdown/WalletDropdown'
