+++ packages/rainbowkit/src/components/RainbowkitThemeProvider/RainbowkitThemeProvider.tsx
@@ -2,17 +2,14 @@ import React, { ReactNode } from 'react'
 import { Theme } from '../../css/sprinkles.css'
 import { cssStringFromTheme } from '../../css/cssStringFromTheme'
 
-export function RainbowkitThemeProvider({
-  id,
-  children,
-  theme,
-  darkModeTheme
-}: {
+export interface RainbowkitThemeProviderProps {
   id?: string
   children: ReactNode
   theme: Theme | (() => Theme)
   darkModeTheme?: Theme | (() => Theme)
-}) {
+}
+
+export function RainbowkitThemeProvider({ id, children, theme, darkModeTheme }: RainbowkitThemeProviderProps) {
   const selector = id ? `[data-rk-id="${id}"]` : '[data-rk]'
 
   const themeCss = `${selector}{${cssStringFromTheme(theme)}}`
