+++ packages/rainbowkit/src/components/Tx/Tx.tsx
@@ -1,8 +1,9 @@
 import { BaseProvider } from '@ethersproject/providers'
 import React, { useEffect, useMemo, useState } from 'react'
 import clsx from 'clsx'
-import { chainIDToExplorer, guessTitle } from '../../utils'
-import type { TransactionWithStatus, TransactionStatus } from '../../hooks'
+import { chainIDToExplorer } from '../../utils/convert'
+import { guessTitle } from '../../utils/guessTitle'
+import type { TransactionWithStatus, TransactionStatus } from '../../hooks/useTxHistory'
 import { FailIcon, LoadingIcon, SuccessIcon, ViewTransactionIcon } from './icons'
 import { Box } from '../Box/Box'
 import { Text } from '../Text/Text'
