+++ packages/rainbowkit/src/components/index.ts
@@ -1,13 +1,25 @@
-export * from './EthAddress/EthAddress'
-export * from './Tx/Tx'
-export * from './AccountInfo/AccountInfo'
-export * from './ExplorerLink/ExplorerLink'
-export * from './CopyAddressButton/CopyAddressButton'
-export * from './TxHistory/TxHistory'
-export * from './NetworkSelect/NetworkSelect'
-export * from './Profile/Profile'
-export * from './WalletDropdown/WalletDropdown'
-export * from './Badge/Badge'
-export * from './EmojiIcon/EmojiIcon'
-export * from './Modal/Modal'
-export * from './RainbowkitThemeProvider/RainbowkitThemeProvider'
+export { AccountInfo } from './AccountInfo/AccountInfo'
+export type { AccountInfoProps } from './AccountInfo/AccountInfo'
+export { Badge } from './Badge/Badge'
+export type { BadgeProps } from './Badge/Badge'
+export { CopyAddressButton } from './CopyAddressButton/CopyAddressButton'
+export { EmojiIcon } from './EmojiIcon/EmojiIcon'
+export type { EmojiIconProps } from './EmojiIcon/EmojiIcon'
+export { EthAddress } from './EthAddress/EthAddress'
+export type { EthAddressProps } from './EthAddress/EthAddress'
+export { ExplorerLink } from './ExplorerLink/ExplorerLink'
+export type { ExplorerLinkProps } from './ExplorerLink/ExplorerLink'
+export { Modal } from './Modal/Modal'
+export type { ModalProps } from './Modal/Modal'
+export { NetworkSelect } from './NetworkSelect/NetworkSelect'
+export type { NetworkSelectProps } from './NetworkSelect/NetworkSelect'
+export { Profile } from './Profile/Profile'
+export type { ProfileProps } from './Profile/Profile'
+export { RainbowkitThemeProvider } from './RainbowkitThemeProvider/RainbowkitThemeProvider'
+export type { RainbowkitThemeProviderProps } from './RainbowkitThemeProvider/RainbowkitThemeProvider'
+export { Tx } from './Tx/Tx'
+export type { TxProps } from './Tx/Tx'
+export { TxHistory } from './TxHistory/TxHistory'
+export type { TxHistoryProps } from './TxHistory/TxHistory'
+export { WalletDropdown } from './WalletDropdown/WalletDropdown'
+export type { WalletDropdownProps } from './WalletDropdown/WalletDropdown'
