--- packages/rainbowkit/src/hooks/index.ts
@@ -1,10 +1,0 @@
-export * from './useExplorerTxHistory'
-export * from './useTxHistory'
-export * from './useWalletInfo'
-export * from './useSignificantBalance'
-export * from './useWeb3State'
-export * from './useSignMessage'
-export * from './useENSWithAvatar'
-export * from './useToggle'
-export * from './useOnClickOutside'
-export * from './useWalletModal'
