--- packages/rainbowkit/src/themes/index.ts
@@ -1,3 +1,0 @@
-export { lightTheme } from './lightTheme'
-export { darkTheme } from './darkTheme'
-export { dimTheme } from './dimTheme'
