--- packages/rainbowkit/src/utils/index.ts
@@ -1,13 +1,0 @@
-export * from './withWeb3React'
-export * from './isAuthorized'
-export * from './convert'
-export * from './address'
-export * from './toSignificant'
-export * from './guessTitle'
-export * from './chains'
-export * from './network'
-export * from './wallets'
-export * from './fetchers'
-export * from './colors'
-export * from './resolveAddress'
-export * from './wc'
